reorg sync utils into time-range-params
