+++ ui/dashboards/src/utils/index.ts
@@ -11,7 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './components';
-export * from './context';
-export * from './views';
-export * from './utils/time-range-params';
+export * from './time-range-params';
